refactor(procmond): split rpc_service.rs into rpc_service/ submodules
Pure mechanical module extraction with no logic or public-API changes:
- rpc_service/error.rs: RpcServiceError + RpcServiceResult
- rpc_service/config.rs: RpcServiceConfig + Default + timeout const
- rpc_service/stats.rs: RpcServiceStats
- rpc_service/tests.rs: full unit test module
- rpc_service/mod.rs: RpcServiceHandler struct + impl, re-exports preserving API

All 51 rpc_service unit tests pass; build/clippy/fmt clean.
